Fixed a bug, normally the level loading must be stable [unstable]

+++ src/level/EnemyData.hpp
@@ -39,7 +39,7 @@ struct EnemyData
     unsigned int hp;            /**< The health points of the enemy */
     unsigned int att;           /**< The attack value of the enemy */
     unsigned int sh;            /**< The defense value of the enemy */
-    uint64_t time;              /**< The time for the enemy to come */
+    unsigned int time;          /**< The time for the enemy to come */
     unsigned int y;             /**< The Y position of the enemy */
     unsigned int w;             /**< The width of the enemy */
     unsigned int h;             /**< The height of the enemy */
